move IncidentsSidebar
